Changed up Balancing and Added Diversity
Added to diversity to the roster of non leviathans available in the B.I.G. Vat. It now includes space amoebas and Tiyanki that roughly fulfill the same combat role as the broodmothers and Swarm warriors they're replacing.

Decreased the single cost of many of the leviathans and switched their expenses much more towards upkeep following a 3% upkeep thats present for many of the vanilla ships