Added scripted trigger for frame relocation exclusion
Currently, adding a planet to class to ignore frame relocation rules (like in the style of the synaptic lathe) requires overwriting two events in giga_210_origins_frameworld.txt.

This is a compatibility issue, and also makes it harder to safely override this.

I've created **giga_is_excluded_from_frameworld_resettlement**, which only contains the synaptic lathe