Update giga_planetcraft_components.txt
Heya, made this change since I found this problem a few days ago and found the way to fix it, I thought I'd share it here '_'. 

This is also my first pull request on a project like this so do be kind if I make a mistake somewhere

Change from ship_modifier to modifier. Currently, ship_modifier applies the effects two times (see bug report https://discord.com/channels/739835273969664050/1415173497562464356 ). This was tested ingame and did indeed solve the problem. I did not look if any other components are in the same situation or not.